+++ app-updater/src/main/java/com/king/app/updater/http/IHttpManager.java
@@ -7,6 +7,7 @@ import java.io.Serializable;
 import java.util.Map;
 
 /**
+ * IHttpManager 默认提供{@link HttpManager} 和 {@link OkHttpManager}两种实现。
  * @author Jenly <a href="mailto:jenly1314@gmail.com">Jenly</a>
  */
 public interface IHttpManager {
@@ -39,7 +40,7 @@ public interface IHttpManager {
          * @param progress
          * @param total
          */
-        void onProgress(int progress,int total);
+        void onProgress(long progress,long total);
 
         /**
          * 完成
